chore: rename user.interface to user.entity
