+++ src/modules/auth/models/AuthRequest.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
-import { User } from 'src/modules/user/user.interface';
+import { User } from 'src/modules/user/user.entity';
 
 export interface AuthRequest extends Request {
   user: User;
